chore: update model name

+++ convert/convert_onnx_quantized.py
@@ -1,12 +1,17 @@
 import argparse
 
+from pathlib import Path
+
+import torch
 from optimum.onnxruntime import ORTModelForCausalLM
 from optimum.onnxruntime.configuration import AutoQuantizationConfig
 from optimum.onnxruntime import ORTQuantizer
-from transformers import AutoTokenizer
+from optimum.exporters.onnx.model_configs import MistralOnnxConfig
+from optimum.exporters.onnx import export, main_export
+from transformers import AutoTokenizer, AutoConfig, AutoModelForCausalLM
 
-MODEL_NAME = "p1atdev/dart-v2-llama-100m"
-SAVE_DIR = "./output/onnx-llama-100m"
+MODEL_NAME = "p1atdev/dart-v2-base"
+SAVE_DIR = "./output/onnx-v2-base"
 
 
 def parse_args():
@@ -31,6 +36,12 @@ def parse_args():
         action="store_true",
         help="Quantize the model",
     )
+    parser.add_argument(
+        "--force",
+        "-f",
+        action="store_true",
+        help="Force convert the model even if the model is not supported",
+    )
     return parser.parse_args()
 
 
@@ -39,19 +50,46 @@ def main():
     model_name = args.model_name
     save_dir = args.save_dir
     do_quantize = args.quantize
+    is_force = args.force
 
-    model = ORTModelForCausalLM.from_pretrained(model_name, export=True)
+    config = AutoConfig.from_pretrained(model_name, trust_remote_code=True)
+    onnx_config = MistralOnnxConfig(config, task="text-generation")
 
-    if do_quantize:
-        qconfig = AutoQuantizationConfig.arm64(is_static=False, per_channel=False)
-        quantizer = ORTQuantizer.from_pretrained(model)
+    if not is_force:
+        model = ORTModelForCausalLM.from_pretrained(model_name, export=True)
 
-        quantizer.quantize(save_dir=save_dir, quantization_config=qconfig)
+        if do_quantize:
+            qconfig = AutoQuantizationConfig.arm64(is_static=False, per_channel=False)
+            quantizer = ORTQuantizer.from_pretrained(model)
 
-        print("Quantization done!")
+            quantizer.quantize(save_dir=save_dir, quantization_config=qconfig)
+
+            print("Quantization done!")
+        else:
+            print("Skipping quantization")
+            model.save_pretrained(save_dir)
+
     else:
-        print("Skipping quantization")
-        model.save_pretrained(save_dir)
+        model = AutoModelForCausalLM.from_pretrained(model_name, trust_remote_code=True)
+
+        export(
+            model,
+            onnx_config,
+            Path(save_dir) / "model.onnx",
+            onnx_config.DEFAULT_ONNX_OPSET,
+        )
+        config.save_pretrained(save_dir)
+
+        if do_quantize:
+            ort_model = ORTModelForCausalLM.from_pretrained(
+                save_dir, export=False, use_cache=False, use_io_binding=False
+            )
+            qconfig = AutoQuantizationConfig.arm64(is_static=False, per_channel=False)
+            quantizer = ORTQuantizer.from_pretrained(ort_model)
+
+            quantizer.quantize(save_dir=save_dir, quantization_config=qconfig)
+
+            print("Quantization done!")
 
 
 if __name__ == "__main__":
